# openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
    title: Norton
    version: 1.0.0
servers:
    -
        url: /api/v1
paths:
...
    '/users/{userId}':
        summary: Path used to manage a single User.
        description: >-
            The REST endpoint/path used to get, update, and delete single instances of an User.  This path
            contains GET, PUT, and DELETE operations used to perform the get, update, and delete tasks,
            respectively.
        get:
            tags:
                - User
            parameters:
                -
                    style: simple
                    explode: false
                    name: userId
                    description: A unique identifier for a User.
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                    description: Successful response - returns a single User.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: get_user
            summary: Get a User
            description: Gets the details of a single instance of a User.
            x-openapi-router-controller: openapi_server.controllers.user_controller
        put:
            requestBody:
                description: Updated User information.
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                required: true
            tags:
                - User
            parameters:
                -
                    style: simple
                    explode: false
                    name: userId
                    description: A unique identifier for a User.
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '202':
                    description: Successful response.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            operationId: update_user
            summary: Update a User
            description: Updates an existing User.
            x-openapi-router-controller: openapi_server.controllers.user_controller
        delete:
            tags:
                - User
            parameters:
                -
                    style: simple
                    explode: false
                    name: userId
                    description: A unique identifier for a User.
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Successful response.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            operationId: delete_user
            summary: Delete a User
            description: Deletes an existing User.
            x-openapi-router-controller: openapi_server.controllers.user_controller
components:
    schemas:
...
        User:
            required:
                - id
                - name
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
            example:
                id: 0
                name: name
...
        Error:
            required:
                - code
                - messages
            type: object
            properties:
                code:
                    type: integer
                messages:
                    type: string
...
    securitySchemes:
        LoginRequired:
            type: apiKey
            name: session_id
            in: cookie
            x-apikeyInfoFunc: openapi_server.controllers.security_controller_.info_from_LoginRequired